Fix a few bugs
Revert removal of jquery-ui reference in html. It was to the css, not js, so needed.
Video variable is not accessible in the tree creation method, so look it up.
When trying to find the section to update on the server, crash on the roots of another medium (parent is not a section and is not the medium we're looking for).