Replace jsTree by jqTree
Simpler to interact with the tree.
Can modify the tree instead of recreating it every time.
Use select (click) and context menu (right click) instead of double click and click.
Not sure about the usability of this one yet, but click/dbclick is problematic because dbclick includes click on several browsers.