Medium deletion behavior
Display a confirmation view instead of the selection view.
Prevent edition and deletion of the "Choose One" element.
Use a promise as a NullObject/Option for getting the selected element.
It fails if the "Choose One" element is selected.
Undefined the selected element in the router if "Choose One" is selected.
Not sure if its useful, but better safe.