More explicit workflow with promises
Views which result is binary (worked or didn't) can return a promise that will be forwarded by the region.
So the calling code can react on the result of the view.
Implemented it for edition and deletion.
Added validation with Backbone.Validation as a trial run.
If we cancel an edition, the model is still updated, so need to fix that.