Bootstrap media data
Application template populate bootstrap data from the db, which avoid 1 unecessary round trip to the server after init.
Stop listening to all media collection events in the media view, otherwise the view is re rendered too often and the medium selection doesn't work.
Stop using Backbone.Validation as it is too complex for current need and not so easy to use.
Validate by hand and show a star with a warning class when title is empty.